:art: Split Yugioh views and urls in respective domains
