+++ source/YugiAcademy/YuGiOh/booster_packs/views/booster_pack_card_view.py
@@ -1,0 +1,22 @@
+from django.shortcuts import redirect, render
+
+from YuGiOh.booster_packs.forms import BoosterPackCardForm
+from YuGiOh.booster_packs import BoosterPackCard
+from YuGiOh.models import booster_pack_system
+
+
+def find_or_store_booster_pack_card(request):
+    if request.method == 'POST':
+        form = BoosterPackCardForm(request.POST)
+        if form.is_valid():
+            booster_pack_card = BoosterPackCard.from_form(form.cleaned_data)
+            booster_pack_system.store_booster_pack_card(booster_pack_card)
+            return redirect(f'/yugioh/booster-pack/{booster_pack_card.booster_pack.id}')
+
+    if request.method == 'GET':
+        context = {
+            'form': BoosterPackCardForm()
+        }
+        return render(request, 'YuGiOh/booster_pack_card_registration.html', context)
+
+    raise Exception(f'The {request.method} method was not expected')
