+++ source/YugiAcademy/YuGiOh/cards/views/__init__.py
@@ -1,0 +1,2 @@
+from .cards_view import cards
+from .spell_card_view import find_or_store_spell_card
