+++ source/YugiAcademy/YuGiOh/models/__init__.py
@@ -1,2 +1,4 @@
 from YuGiOh.cards import Card, MonsterCard, SpellCard, TrapCard
 from YuGiOh.booster_packs import BoosterPack, BoosterPackCard
+from .system_aware_contexts import card_system, booster_pack_system
+
