+++ source/YugiAcademy/YuGiOh/website/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from YuGiOh.website.views import home, about
+
+urlpatterns = [
+    path('', home, name="home"),
+    path('about', about, name="about")
+]
