+++ source/YugiAcademy/YuGiOh/website/views/__init__.py
@@ -1,0 +1,2 @@
+from .home_view import home
+from .about_view import about
