+++ source/YugiAcademy/YuGiOh/website/views/about_view.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def about(request):
+    return render(request, "about.html")
