+++ source/YugiAcademy/YuGiOh/website/views/home_view.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def home(request):
+    return render(request, "index.html")
